@@ -1,5 +1,5 @@
 module github.com/vimeo/go-status-page
 
-go 1.20
+go 1.24.0
 
-require golang.org/x/net v0.12.0
+require golang.org/x/net v0.46.0
